pyproject.toml
[tool.poetry]
name = "auto-sender"
version = "0.1.0"
description = ""
authors = ["piotek8 <piotr.bator.contact@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
selenium = "^4.16.0"
webdriver-manager = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
